refactor(weather): move `getSunriseSunsetTime` to RSC instead of server action
